Split ExtentsBlocks into a new submodule of iters::file_blocks
The BlockMap variant of FileBlocks will be fairly complicated, so I
intend to put it in its own module. To keep the code organization
consistent, move ExtentsBlocks into its own module too.
